MOved helium.modes under backend.models (#82)
Commented out WebSocket stuff
