+++ src/main/java/com/wire/kalium/backend/models/Access.kt
@@ -1,0 +1,11 @@
+package com.wire.kalium.backend.models
+
+import java.util.*
+
+data class Access(
+        val cookie: Cookie,
+        val userId: UUID,
+        val access_token: String,
+        val expires_in: Int,
+        val token_type: String,
+)
