+++ src/main/java/com/wire/kalium/backend/models/NotificationList.kt
@@ -1,0 +1,6 @@
+package com.wire.kalium.backend.models
+
+data class NotificationList(
+        val has_more: Boolean,
+        val notifications: MutableList<Event>
+)
